stage2_conditional/requirements.txt
# Stage 2: Qwen2-VL 7B Requirements

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.25.0

# Quantization (for later)
bitsandbytes>=0.41.0

# Fine-Tuning
peft>=0.7.0  # For LoRA/QLoRA

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
scikit-learn>=1.3.0
pyyaml>=6.0

# YOLO Integration (for Stage 1)
ultralytics>=8.0.0

# Evaluation
scipy>=1.11.0


